better inscription inlining
